+++ lib/ts/recipe/session/recipe.ts
@@ -78,7 +78,11 @@ export default class SessionRecipe extends RecipeModule {
             });
 
             let builder = new OverrideableBuilder(
-                RecipeImplementation(Querier.getNewInstanceOrThrowError(recipeId), this.config)
+                RecipeImplementation(
+                    Querier.getNewInstanceOrThrowError(recipeId),
+                    this.config,
+                    () => this.recipeInterfaceImpl
+                )
             );
             this.recipeInterfaceImpl = builder
                 .override((oI) => {
@@ -94,7 +98,11 @@ export default class SessionRecipe extends RecipeModule {
         } else {
             {
                 let builder = new OverrideableBuilder(
-                    RecipeImplementation(Querier.getNewInstanceOrThrowError(recipeId), this.config)
+                    RecipeImplementation(
+                        Querier.getNewInstanceOrThrowError(recipeId),
+                        this.config,
+                        () => this.recipeInterfaceImpl
+                    )
                 );
                 this.recipeInterfaceImpl = builder.override(this.config.override.functions).build();
             }